Convert receiver entry point to ASP.NET Core host
